+++ pycvrplib/_parse_utils.py
@@ -1,14 +1,13 @@
 from __future__ import annotations
 
-import inspect
-import math
 import re
 from collections import defaultdict
 from dataclasses import dataclass
 from itertools import combinations
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Optional
 
 from .constants import DEPOT, SHIFT
+from .utils import euclidean, from_dict_to_dataclass
 
 """
 Instance utilities
@@ -124,30 +123,6 @@ def get_representation(edge_weights: List[List[int]], n: int) -> str:
         return "flattened"
 
 
-def euclidean(coords: List[List[int]], round_func=round) -> List[List[int]]:
-    """
-    Compute the pairwise Euclidean distances using the passed-in coordinates.
-    `round_func` can be used to specify the rouding function. Default is to
-    round to the nearest integer.
-    """
-
-    def dist(p, q):
-        """
-        Return the Euclidean distance between to coordinates.
-        """
-        return math.sqrt(sum((px - qx) ** 2.0 for px, qx in zip(p, q)))
-
-    n = len(coords)
-    distances = [[0 for _ in range(n)] for _ in range(n)]
-
-    for (i, coord_i), (j, coord_j) in combinations(enumerate(coords), r=2):
-        d_ij = round_func(dist(coord_i, coord_j))
-        distances[i][j] = d_ij
-        distances[j][i] = d_ij
-
-    return distances
-
-
 def from_triangular(triangular: List[List[int]]) -> List[List[int]]:
     """
     Compute a full distances matrix from a triangular matrix.
@@ -180,15 +155,6 @@ def from_flattened(edge_weights: List[List[int]], n: int) -> List[List[int]]:
         distances[j][i] = d_ij
 
     return distances
-
-
-def from_dict_to_dataclass(cls, data):
-    return cls(
-        **{
-            key: (data[key] if val.default == val.empty else data.get(key, val.default))
-            for key, val in inspect.signature(cls).parameters.items()
-        }
-    )
 
 
 @dataclass
